+++ lib/plugin_config.js
@@ -1,12 +1,12 @@
 var resolve = require('path').resolve;
 var readFileSync = require('fs').readFileSync;
-var fileConfig = require('./file_config');
+var configFile = require('./config_file');
 
 module.exports = function (root) {
   if (!root) root = process.cwd();
 
   var pkg = require(resolve(root, 'package.json'));
-  var config = fileConfig(root);
+  var config = configFile(root);
 
   var buildSourcePatterns = [
     'package.json',
